+++ Core/Src/main.c
@@ -104,6 +104,8 @@ int main(void)
   MX_USART1_UART_Init();
   MX_USART10_UART_Init();
   MX_TIM4_Init();
+  MX_USART2_UART_Init();
+  MX_USART3_UART_Init();
   /* USER CODE BEGIN 2 */
 	Task_Init();